+++ documentation/source/CHANGELOG.md
@@ -1,7 +1,11 @@
+# 0.3.0
+
+* onBeforeChange now receives a segment object as the first argument.
+* onAfterChange now receives a segment object as the first argument.
+
 # 0.2.5
 
 * Removed instance.firstRun variable.
-
 
 # 0.2 2018-11-29T07:45